chore: remove redundant cast in MqttClientConfigBuilder.LimitUnacknowledgedPublications

+++ src/Silverback.Integration.MQTT/Messaging/Configuration/Mqtt/MqttClientConfigBuilder.cs
@@ -287,7 +287,7 @@ namespace Silverback.Messaging.Configuration.Mqtt
         /// <inheritdoc cref="IMqttClientConfigBuilder.LimitUnacknowledgedPublications" />
         public IMqttClientConfigBuilder LimitUnacknowledgedPublications(ushort receiveMaximum)
         {
-            _builder.WithReceiveMaximum((ushort)receiveMaximum);
+            _builder.WithReceiveMaximum(receiveMaximum);
             return this;
         }
